refactor(ios): co-locate LoginViewModel in LoginView.swift
Move the 32-line LoginViewModel into LoginView.swift following the
architecture rule that ViewModels live in the same file as their view.
